>Cambios en tipos de dato

+++ APPWEB-INCOMEMotorCalculoJava-api/src/main/java/com/inditex/rrhh/icmclcwb/api/app/periodo/dto/EstadoPeriodoCalculoPersonaEnum.java
@@ -7,12 +7,12 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public enum EstadoPeriodoCalculoPersonaEnum {
 
-    SIN_CALCULAR(1L, new EstadoPeriodoCalculoPersonaDto(1L)), CALCULADO(2L, new EstadoPeriodoCalculoPersonaDto(2L)),
-    CONFIRMADO(3L, new EstadoPeriodoCalculoPersonaDto(3L)), CALCULADO_ERROR(4L, new EstadoPeriodoCalculoPersonaDto(4L)),
-    EXPORTADO(5L, new EstadoPeriodoCalculoPersonaDto(5L)), RECUPERADO(6L, new EstadoPeriodoCalculoPersonaDto(6L)),
-    PAGADO(7L, new EstadoPeriodoCalculoPersonaDto(7L));
+    SIN_CALCULAR(1, new EstadoPeriodoCalculoPersonaDto(1)), CALCULADO(2, new EstadoPeriodoCalculoPersonaDto(2)),
+    CONFIRMADO(3, new EstadoPeriodoCalculoPersonaDto(3)), CALCULADO_ERROR(4, new EstadoPeriodoCalculoPersonaDto(4)),
+    EXPORTADO(5, new EstadoPeriodoCalculoPersonaDto(5)), RECUPERADO(6, new EstadoPeriodoCalculoPersonaDto(6)),
+    PAGADO(7, new EstadoPeriodoCalculoPersonaDto(7));
 
-    private final Long id;
+    private final Integer id;
 
     private final EstadoPeriodoCalculoPersonaDto dto;
