+++ APPWEB-INCOMEMotorCalculoJava-model/src/main/java/com/inditex/rrhh/icmclcwb/model/primary/tarea/repository/TareaPersonaHistoricoRepositoryCustomImpl.java
@@ -99,7 +99,7 @@ public class TareaPersonaHistoricoRepositoryCustomImpl
     
     @Override
     public List<IdPersonaHistoricoDto> findIdPersonaHistoricoDtoByIdTareaAndIdOrigenAndTipoDatoInAmbito(@NotNull @Positive final Long idTarea, @NotBlank final String idOrigen,
-            @NotNull final List<Long> idsTipoDato) {
+            @NotNull final List<Integer> idsTipoDato) {
         MapSqlParameterSource parameters = new MapSqlParameterSource();
         parameters.addValue(SqlPrimaryConstants.SQL_PARAM_ID_TAREA, idTarea);
         parameters.addValue(SqlPrimaryConstants.SQL_PARAM_ID_ORIGEN, idOrigen);
